Terminal Fix, minor changes - code update
I lost multiple days worth of code due to a git commit (several
actually) that caused me some consternation. Now I have to catch back
up on those changes. This is a step in that direction. I’ll need to
re-add the humanoid character, and re-layout the new terminal window.