Created new industrial windmill feature in Blender. Character controller still hosed.  :-(
The first of a number of landscape atmospheres that I’ve created to
include in the game. The character controller is still broken. Need to
work on network chat, get a working controller, create a shoot em up
ability with full score and health tracking. Also need to integrate the
new 2D port scanner and other C# features.  Loads to do!